Refactoring towards simplicity
