+++ src/components/general/landing.vue
@@ -1,9 +1,9 @@
 <template>
   <v-card>
-    <v-card-title class="headline justify-center">Intro To Cryptocurrency</v-card-title>
+    <v-card-title class="headline justify-center">{{landingInfo.title}}</v-card-title>
     <v-divider/>
     <v-row no-gutters justify='space-around'>
-       <v-col :key="item.title" v-for="(item) in lessons" cols='4'>
+       <v-col :key="item.title" v-for="(item) in landingInfo.lessons" cols='4'>
          <v-card
            class="ma-3" :color="cardColor(item.unlocked)"
          >
@@ -20,8 +20,8 @@
         </v-row>
          <v-card-actions class="justify-center">
            <v-btn
-             color="deep-purple accent-4"
-             :disabled='!unlocked(item.unlocked)'
+             :color="iconColor(true)"
+             :disabled='!item.unlocked'
              @click="startLesson (item.comp)"
            >
              Start
@@ -51,41 +51,25 @@
 </template>
 <script>
 export default {
+  props: ['landingInfo'],
   data: () => ({
-    componentList: {
-    },
-    lesson: '',
-    lessons: [
-      { title: 'What is Cryptocurrency?', icon: 'atom-variant', comp: 'whyCrypto', unlocked: 'true' },
-      { title: 'Which Cryptocurrency is Best For Me?', icon: 'account-question', comp: 'whichCrypto', unlocked: 'true' }
-      // { title: 'Pick Your First Cryptocurrency', icon: 'account-cash', unlocked: 'false' }
-    ]
   }),
   components: {
   },
   computed: {
-    currentMain () {
-      return () => import('@/components/edu/intro/landing.vue')
-    }
   },
   methods: {
     cardColor (state) {
-      if (state === 'true') {
+      if (state) {
         return 'blue'
       }
-      return 'grey darken-4'
+      return 'grey darken-3'
     },
     iconColor (state) {
-      if (state === 'true') {
-        return 'black'
+      if (state) {
+        return 'indigo darken-4'
       }
       return 'white'
-    },
-    unlocked (state) {
-      if (state === 'true') {
-        return true
-      }
-      return false
     },
     exit () {
       this.$emit('exit')
